+++ dashboard/lib/module_orders/manager/orders_manager/orders_manager.dart
@@ -81,6 +81,7 @@ class OrdersManager {
   Future<ActionResponse?> updateAnswerOrderCashForStore(
           StoreAnswerForOrderCashRequest request) =>
       _repository.updateAnswerOrderCashForStore(request);
-  Future<ActionResponse?> recycleOrderStatus(UpdateOrderRequest request) =>
-      _repository.recycleOrderStatus(request);
+
+  Future<ActionResponse?> recycleOrder(CreateOrderRequest request) =>
+      _repository.recycleOrder(request);
 }
